+++ src/main/java/org/apache/commons/lang3/mutable/MutableFloat.java
@@ -84,7 +84,7 @@ public class MutableFloat extends Number implements Comparable<MutableFloat>, Mu
      */
     @Override
     public Float getValue() {
-        return this.value;
+        return Float.valueOf(this.value);
     }
 
     /**
@@ -343,7 +343,7 @@ public class MutableFloat extends Number implements Comparable<MutableFloat>, Mu
      * @return a Float instance containing the value from this mutable, never null
      */
     public Float toFloat() {
-        return floatValue();
+        return Float.valueOf(floatValue());
     }
 
     //-----------------------------------------------------------------------
